+++ euler579c/util.cpp
@@ -19,7 +19,7 @@ long long util::gcd_l(long long a, long long b)
 
 long long util::gcd(set<long long>::iterator first, set<long long>::iterator last)
 {
-	long l1 = *first;
+	long long l1 = *first;
 	if (++first == last) return l1;
 	else return gcd_l(abs(l1), abs(gcd(first, last)));
 }
